automation/pyproject.toml
[project]
name = "nifi-rest-automation"
version = "0.1.0"
description = "CLI and library for provisioning Apache NiFi flows via the REST API."
readme = "README.md"
authors = [
    { name = "NiFi Automation Toolkit" }
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "httpx>=0.27,<0.28",
    "typer>=0.12,<0.13",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "rich>=13.7,<14.0",
    "pyyaml>=6.0,<7.0"
]

[project.scripts]
nifi-automation = "nifi_automation.cli:app"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.6",
  "mypy>=1.10"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
markers = [
  "integration: marks tests that require a live NiFi instance",
  "tools: marks tests for tools CLI flows (trust-store ops)"
]
